scripts/requirements.txt
pyannote.audio>=3.1
torch
soundfile
numpy
requests
transformers>=4.56,!=5.0.*,!=5.1.*
accelerate
